Make portfolio the index route
